[ISSUE #9619]♻️Extract RemotingCommand inline tests into a sibling module
